pyproject.toml
[project]
name = "sirius-compass-backend"
version = "0.1.0"
description = "Agentic AI platform to analyze projects and provide status."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "Proprietary"}
authors = [
    {name = "Federico Martucci", email = "federicomartucci@sirius.com.ar"},
    {name = "Santiago Alvarez", email = "santiagoalvarez@sirius.com.ar"}
]
keywords = ["ai", "agents", "engineering-metrics", "rag", "hexagonal-architecture"]
dependencies = [
    # Utilities
    "python-dotenv>=1.2.1",       # Environment variables management (.env)
    "pydantic>=2.12.5",            # Data validation and schemas (Domain Models)
    "requests>=2.32.5",           # Generic HTTP requests

    # Core & Agents
    "langchain>=1.1.3",               # Base framework
    "langgraph>=1.0.4",              # Agent orchestration (Graphs)
    "langchain-openai>=1.1.1",        # Connector for GPT-4o (optional/future)
    "langchain-google-genai>=4.0.0",  # Connector for Gemini (current choice)

    # Vector Database (RAG)
    "langchain-pinecone>=0.2.13",      # LangChain integration with Pinecone

    # Data Analytics & Visualization
    "pandas>=2.3.3",              # Tabular data manipulation (Reports)
    "matplotlib>=3.10.7",          # Static chart generation (images in chat)
    "plotly>=6.5.0",             # Interactive charts

    # Adapters
    "PyGithub>=2.8.1",            # SDK to adapt GitHub
    # "py-trello"

    # Web Service
    "fastapi>=0.124.0",
    "uvicorn>=0.38.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Configuration to automatically discover source code
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.poetry]
package-mode = false
